auth page metadata
